added new features to safechat and added some permissions

+++ src/safeChat/src/options/closeChat.js
@@ -1,21 +1,36 @@
-const closeChat = async (message, args, server, safeChatUsers) => {
-    const getSafeChatUser = await safeChatUsers.findOne({where: {dc_UserId: message.author.id }}).then((safeChatUser) => {
-        let getServerChannel = server.channels.cache.get(safeChatUser.dc_channelId)
-        if(safeChatUser){
-            
-            if (getServerChannel) {
-                getServerChannel.delete().then(() => {
-                    safeChatUser.destroy()
-                    message.channel.send("Safe chat session has ended!")
-                })
+const closeChat = async (message, args, server, safeChatUsers, client) => {
+    if (message.channel.type == "DM") {
+        const getSafeChatUser = await safeChatUsers.findOne({where: {dc_UserId: message.author.id }}).then((safeChatUser) => {
+            if(safeChatUser){
+                let getServerChannel = server.channels.cache.get(safeChatUser.dc_channelId)
+                if (getServerChannel) {
+                    getServerChannel.send("<@" + safeChatUser.dc_staffUserId + ">, The resident wants to close this session, use `+closechat` to end this session.")
+                    message.reply("Our moderators have been notified that you want to close this session! Please be patient while we close your session!\n\n**After closing this session, all chat history will be automatically deleted**")
+                } else {
+                    message.channel.send("Can't close session channel doesn't exist!")
+                }
+                
             } else {
-                message.channel.send("Can't close session channel doesn't exist! Contact a staff member!")
+                message.channel.send("You don't have an active safe chat session!")
             }
-            
-        } else {
-            message.channel.send("You don't have an active safe chat session!")
-        }
-    })
+        })
+    } else {
+        const getSafeChatUser = await safeChatUsers.findOne({where: {dc_staffUserId: message.author.id }}).then((safeChatUser) => {
+            if(safeChatUser){
+                let getServerChannel = server.channels.cache.get(safeChatUser.dc_channelId)
+                let getUser = client.users.cache.find(user => user.id === safeChatUser.dc_UserId)
+                if (getServerChannel) {
+                    getServerChannel.delete().then(() => {
+                        safeChatUser.destroy()
+                        getUser.send("Safe chat session has been closed by the moderator")
+                    })
+                } else {
+                    message.channel.send("Can't close session channel doesn't exist!")
+                }
+            }
+        })
+    }
+
 }
 
 module.exports = { closeChat }
